refactor(tests): fixed renaming of files and ported over redirect tests to public form controller
