+++ src/app/modules/auth/__tests__/auth.controller.spec.ts
@@ -1,23 +1,15 @@
-import { ObjectId } from 'bson-ext'
-import { err, errAsync, ok, okAsync } from 'neverthrow'
+import { errAsync, okAsync } from 'neverthrow'
 import { mocked } from 'ts-jest/utils'
 
-import { MissingFeatureError } from 'src/app/modules/core/core.errors'
 import MailService from 'src/app/services/mail/mail.service'
 import { HashingError } from 'src/app/utils/hash'
-import { AuthType, IAgencySchema, IUserSchema } from 'src/types'
+import { IAgencySchema, IUserSchema } from 'src/types'
 
 import dbHandler from 'tests/unit/backend/helpers/jest-db'
 import expressHandler from 'tests/unit/backend/helpers/jest-express'
 
 import { MailSendError } from '../../../services/mail/mail.errors'
 import { DatabaseError } from '../../core/core.errors'
-import * as FormService from '../../form/form.service'
-import { MyInfoFactory } from '../../myinfo/myinfo.factory'
-import { IMyInfoForm } from '../../myinfo/myinfo.types'
-import { CreateRedirectUrlError } from '../../spcp/spcp.errors'
-import { SpcpFactory } from '../../spcp/spcp.factory'
-import { ISPCPForm } from '../../spcp/spcp.types'
 import * as UserService from '../../user/user.service'
 import * as AuthController from '../auth.controller'
 import { InvalidDomainError, InvalidOtpError } from '../auth.errors'
@@ -35,9 +27,6 @@ jest.mock('../../myinfo/myinfo.factory')
 const MockAuthService = mocked(AuthService)
 const MockMailService = mocked(MailService)
 const MockUserService = mocked(UserService)
-const MockFormService = mocked(FormService)
-const MockSpcpFactory = mocked(SpcpFactory, true)
-const MockMyInfoFactory = mocked(MyInfoFactory, true)
 
 describe('auth.controller', () => {
   beforeAll(async () => await dbHandler.connect())
@@ -360,308 +349,6 @@ describe('auth.controller', () => {
       expect(mockRes.json).toBeCalledWith({ message: 'Sign out failed' })
       expect(mockDestroyWithErr).toBeCalledTimes(1)
       expect(mockClearCookie).not.toBeCalled()
-    })
-  })
-
-  describe('getRedirectLink', () => {
-    const MOCK_REQ = expressHandler.mockRequest({
-      params: {
-        formId: new ObjectId().toHexString(),
-      },
-      query: {
-        isPersistentLogin: true,
-      },
-    })
-    const MOCK_REDIRECT_URL = 'www.mockata.com'
-
-    it('should return 200 with the redirect url when the request is valid and the form has authType SP', async () => {
-      // Arrange
-      const MOCK_FORM = {
-        authType: AuthType.SP,
-        esrvcId: '12345',
-      } as ISPCPForm
-      const mockRes = expressHandler.mockResponse()
-      MockFormService.retrieveFullFormById.mockReturnValueOnce(
-        okAsync(MOCK_FORM),
-      )
-      MockSpcpFactory.createRedirectUrl.mockReturnValueOnce(
-        ok(MOCK_REDIRECT_URL),
-      )
-
-      // Act
-      await AuthController.getRedirectLink(MOCK_REQ, mockRes, jest.fn())
-
-      // Assert
-      expect(mockRes.status).toBeCalledWith(200)
-      expect(mockRes.json).toBeCalledWith({ redirectURL: MOCK_REDIRECT_URL })
-    })
-
-    it('should return 200 with the redirect url when the request is valid and the form has authType CP', async () => {
-      // Arrange
-      const MOCK_FORM = {
-        authType: AuthType.CP,
-        esrvcId: '12345',
-      } as ISPCPForm
-      const MOCK_REQ = expressHandler.mockRequest({
-        params: {
-          formId: new ObjectId().toHexString(),
-        },
-        query: {
-          isPersistentLogin: true,
-        },
-      })
-      const mockRes = expressHandler.mockResponse()
-      MockFormService.retrieveFullFormById.mockReturnValueOnce(
-        okAsync(MOCK_FORM),
-      )
-      MockSpcpFactory.createRedirectUrl.mockReturnValueOnce(
-        ok(MOCK_REDIRECT_URL),
-      )
-
-      // Act
-      await AuthController.getRedirectLink(MOCK_REQ, mockRes, jest.fn())
-
-      // Assert
-      expect(mockRes.status).toBeCalledWith(200)
-      expect(mockRes.json).toBeCalledWith({ redirectURL: MOCK_REDIRECT_URL })
-    })
-
-    it('should return 200 with the redirect url when the request is valid and the form has authType MyInfo', async () => {
-      // Arrange
-      const MOCK_FORM = ({
-        authType: AuthType.MyInfo,
-        esrvcId: '12345',
-        getUniqueMyInfoAttrs: jest.fn().mockReturnValue([]),
-      } as unknown) as IMyInfoForm
-      const MOCK_REQ = expressHandler.mockRequest({
-        params: {
-          formId: new ObjectId().toHexString(),
-        },
-        query: {
-          isPersistentLogin: true,
-        },
-      })
-      const mockRes = expressHandler.mockResponse()
-      MockFormService.retrieveFullFormById.mockReturnValueOnce(
-        okAsync(MOCK_FORM),
-      )
-      MockMyInfoFactory.createRedirectURL.mockReturnValueOnce(
-        ok(MOCK_REDIRECT_URL),
-      )
-
-      // Act
-      await AuthController.getRedirectLink(MOCK_REQ, mockRes, jest.fn())
-
-      // Assert
-      expect(mockRes.status).toBeCalledWith(200)
-      expect(mockRes.json).toBeCalledWith({ redirectURL: MOCK_REDIRECT_URL })
-    })
-
-    it('should return 400 when the form has authType NIL', async () => {
-      // Arrange
-      const MOCK_FORM = {
-        authType: AuthType.NIL,
-        esrvcId: '12345',
-      }
-      const MOCK_REQ = expressHandler.mockRequest({
-        params: {
-          formId: new ObjectId().toHexString(),
-        },
-        query: {
-          isPersistentLogin: true,
-        },
-      })
-      const mockRes = expressHandler.mockResponse()
-      MockFormService.retrieveFullFormById.mockReturnValueOnce(
-        okAsync(MOCK_FORM),
-      )
-
-      // Act
-      await AuthController.getRedirectLink(MOCK_REQ, mockRes, jest.fn())
-
-      // Assert
-      expect(mockRes.status).toBeCalledWith(400)
-      expect(mockRes.json).toBeCalledWith({
-        message:
-          'Please ensure that the form has authentication enabled. Please refresh and try again.',
-      })
-    })
-
-    it('should return 400 when the form has authType MyInfo and is missing esrvcId', async () => {
-      // Arrange
-      const MOCK_FORM = ({
-        authType: AuthType.MyInfo,
-      } as unknown) as IMyInfoForm
-      const MOCK_REQ = expressHandler.mockRequest({
-        params: {
-          formId: new ObjectId().toHexString(),
-        },
-        query: {
-          isPersistentLogin: true,
-        },
-      })
-      const mockRes = expressHandler.mockResponse()
-      MockFormService.retrieveFullFormById.mockReturnValueOnce(
-        okAsync(MOCK_FORM),
-      )
-
-      // Act
-      await AuthController.getRedirectLink(MOCK_REQ, mockRes, jest.fn())
-
-      // Assert
-      expect(mockRes.status).toBeCalledWith(400)
-      expect(mockRes.json).toBeCalledWith({
-        message:
-          'This form does not have MyInfo enabled. Please refresh and try again.',
-      })
-    })
-
-    it('should return 400 when the form has authType SP and is missing esrvcId', async () => {
-      // Arrange
-      const MOCK_FORM = ({
-        authType: AuthType.SP,
-      } as unknown) as ISPCPForm
-      const MOCK_REQ = expressHandler.mockRequest({
-        params: {
-          formId: new ObjectId().toHexString(),
-        },
-        query: {
-          isPersistentLogin: true,
-        },
-      })
-      const mockRes = expressHandler.mockResponse()
-      MockFormService.retrieveFullFormById.mockReturnValueOnce(
-        okAsync(MOCK_FORM),
-      )
-
-      // Act
-      await AuthController.getRedirectLink(MOCK_REQ, mockRes, jest.fn())
-
-      // Assert
-      expect(mockRes.status).toBeCalledWith(400)
-      expect(mockRes.json).toBeCalledWith({
-        message:
-          'This form does not have Singpass or Corppass enabled. Please refresh and try again.',
-      })
-    })
-
-    it('should return 400 when the form has authType CP and is missing esrvcId', async () => {
-      // Arrange
-      const MOCK_FORM = ({
-        authType: AuthType.CP,
-      } as unknown) as ISPCPForm
-      const MOCK_REQ = expressHandler.mockRequest({
-        params: {
-          formId: new ObjectId().toHexString(),
-        },
-        query: {
-          isPersistentLogin: true,
-        },
-      })
-      const mockRes = expressHandler.mockResponse()
-      MockFormService.retrieveFullFormById.mockReturnValueOnce(
-        okAsync(MOCK_FORM),
-      )
-
-      // Act
-      await AuthController.getRedirectLink(MOCK_REQ, mockRes, jest.fn())
-
-      // Assert
-      expect(mockRes.status).toBeCalledWith(400)
-      expect(mockRes.json).toBeCalledWith({
-        message:
-          'This form does not have Singpass or Corppass enabled. Please refresh and try again.',
-      })
-    })
-
-    it('should return 500 when the form could not be retrieved from the database', async () => {
-      // Arrange
-      const MOCK_REQ = expressHandler.mockRequest({
-        params: {
-          formId: new ObjectId().toHexString(),
-        },
-        query: {
-          isPersistentLogin: true,
-        },
-      })
-      const mockRes = expressHandler.mockResponse()
-      MockFormService.retrieveFullFormById.mockReturnValueOnce(
-        errAsync(new DatabaseError()),
-      )
-
-      // Act
-      await AuthController.getRedirectLink(MOCK_REQ, mockRes, jest.fn())
-
-      // Assert
-      expect(mockRes.status).toBeCalledWith(500)
-      expect(mockRes.json).toBeCalledWith({
-        message: 'Sorry, something went wrong. Please try again.',
-      })
-    })
-
-    it('should return 500 when the redirectURL could not be created', async () => {
-      // Arrange
-      const MOCK_FORM = ({
-        esrvcId: '234',
-        authType: AuthType.CP,
-      } as unknown) as ISPCPForm
-      const MOCK_REQ = expressHandler.mockRequest({
-        params: {
-          formId: new ObjectId().toHexString(),
-        },
-        query: {
-          isPersistentLogin: true,
-        },
-      })
-      const mockRes = expressHandler.mockResponse()
-      MockFormService.retrieveFullFormById.mockReturnValueOnce(
-        okAsync(MOCK_FORM),
-      )
-      MockSpcpFactory.createRedirectUrl.mockReturnValueOnce(
-        err(new CreateRedirectUrlError()),
-      )
-
-      // Act
-      await AuthController.getRedirectLink(MOCK_REQ, mockRes, jest.fn())
-
-      // Assert
-      expect(mockRes.status).toBeCalledWith(500)
-      expect(mockRes.json).toBeCalledWith({
-        message: 'Sorry, something went wrong. Please try again.',
-      })
-    })
-
-    it('should return 500 when the redirectURL feature is not implemented', async () => {
-      // Arrange
-      const MOCK_FORM = ({
-        esrvcId: '234',
-        authType: AuthType.MyInfo,
-        getUniqueMyInfoAttrs: jest.fn().mockReturnValue([]),
-      } as unknown) as ISPCPForm
-      const MOCK_REQ = expressHandler.mockRequest({
-        params: {
-          formId: new ObjectId().toHexString(),
-        },
-        query: {
-          isPersistentLogin: true,
-        },
-      })
-      const mockRes = expressHandler.mockResponse()
-      MockFormService.retrieveFullFormById.mockReturnValueOnce(
-        okAsync(MOCK_FORM),
-      )
-      MockMyInfoFactory.createRedirectURL.mockReturnValueOnce(
-        err(new MissingFeatureError()),
-      )
-
-      // Act
-      await AuthController.getRedirectLink(MOCK_REQ, mockRes, jest.fn())
-
-      // Assert
-      expect(mockRes.status).toBeCalledWith(500)
-      expect(mockRes.json).toBeCalledWith({
-        message: 'Sorry, something went wrong. Please try again.',
-      })
     })
   })
 })
