+++ src/public/services/__tests__/PublicFormAuthService.test.ts
@@ -2,7 +2,7 @@ import axios from 'axios'
 import { ObjectId } from 'bson'
 import { mocked } from 'ts-jest/utils'
 
-import * as AuthService from '../AuthService'
+import * as PublicFormAuthService from '../PublicFormAuthService'
 
 jest.mock('axios')
 const MockAxios = mocked(axios, true)
@@ -10,17 +10,20 @@ const MockAxios = mocked(axios, true)
 const MOCK_REDIRECT_URL = 'redirectURL'
 const MOCK_FORM_ID = new ObjectId().toHexString()
 
-describe('AuthService', () => {
+describe('PublicFormAuthService', () => {
   afterEach(() => jest.resetAllMocks())
   describe('createRedirectURL', () => {
     it('should return the redirect URL when retrieval succeeds and persistent login is not set', async () => {
       const mockData = { redirectURL: MOCK_REDIRECT_URL }
       MockAxios.get.mockResolvedValueOnce({ data: mockData })
 
-      const result = await AuthService.createRedirectURL(MOCK_FORM_ID, false)
+      const result = await PublicFormAuthService.createRedirectURL(
+        MOCK_FORM_ID,
+        false,
+      )
 
       expect(MockAxios.get).toHaveBeenCalledWith(
-        `${AuthService.PUBLIC_FORMS_ENDPOINT}/${MOCK_FORM_ID}/${AuthService.REDIRECT_URL_ENDPOINT}`,
+        `${PublicFormAuthService.PUBLIC_FORMS_ENDPOINT}/${MOCK_FORM_ID}/auth/redirect`,
         {
           params: { isPersistentLogin: false },
         },
@@ -32,10 +35,13 @@ describe('AuthService', () => {
       const mockData = { redirectURL: MOCK_REDIRECT_URL }
       MockAxios.get.mockResolvedValueOnce({ data: mockData })
 
-      const result = await AuthService.createRedirectURL(MOCK_FORM_ID, true)
+      const result = await PublicFormAuthService.createRedirectURL(
+        MOCK_FORM_ID,
+        true,
+      )
 
       expect(MockAxios.get).toHaveBeenCalledWith(
-        `${AuthService.PUBLIC_FORMS_ENDPOINT}/${MOCK_FORM_ID}/${AuthService.REDIRECT_URL_ENDPOINT}`,
+        `${PublicFormAuthService.PUBLIC_FORMS_ENDPOINT}/${MOCK_FORM_ID}/auth/redirect`,
         {
           params: { isPersistentLogin: true },
         },
@@ -47,11 +53,12 @@ describe('AuthService', () => {
       const error = new Error('rejected')
       MockAxios.get.mockRejectedValueOnce(error)
 
-      const rejectFunction = () => AuthService.createRedirectURL(MOCK_FORM_ID)
+      const rejectFunction = () =>
+        PublicFormAuthService.createRedirectURL(MOCK_FORM_ID)
 
       await expect(rejectFunction).rejects.toThrowError(error)
       expect(MockAxios.get).toHaveBeenCalledWith(
-        `${AuthService.PUBLIC_FORMS_ENDPOINT}/${MOCK_FORM_ID}/${AuthService.REDIRECT_URL_ENDPOINT}`,
+        `${PublicFormAuthService.PUBLIC_FORMS_ENDPOINT}/${MOCK_FORM_ID}/auth/redirect`,
         {
           params: { isPersistentLogin: false },
         },
